requirements.txt
asgiref==3.7.2
Django==4.2.9
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
environ==1.0
importlib-metadata==7.0.1
inflection==0.5.1
Markdown==3.5.1
packaging==23.2
pillow==10.2.0
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
shortuuid==1.0.11
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
zipp==3.17.0
